fix: Extract lightweight URL parsing to avoid Temporal sandbox restrictions
The Temporal workflow sandbox restricts imports like httpx and urllib.request.
The workflow was importing tarball_validation.py which transitively imported
fastapi, sqlalchemy, and httpx through its dependency chain.

This fix:
1. Creates tarball_url.py with pure URL parsing functions (no heavy deps)
2. Updates workflows.py to import from the lightweight module
3. Updates tarball_validation.py to re-export from tarball_url.py

This allows the workflow code to use is_http_url() and parse_internal_upload_id()
without triggering sandbox restriction errors.
